requirements.txt
rdkit-pypi
pandas
tqdm
# optional, if you want GPU acceleration locally:
# cudf-cu11  # or appropriate cuDF package for your CUDA toolchain
